This implements the summary message and closes #206
Some methods are moved by Intellij reformat code tool!

New maps for keep the current coins and experience earned until game end. It's unnecessary create classes for this (I guess).
These maps are loaded on player join and destroyed on player leave.
These maps are updated on addXp method and MoneyListeners methods.
The message is implemented below of top killers message.
The radical implements of "long progress bar" is for use in chat. The shorter progress bar it looks very ugly in the rewards summary message. This could be improved or customized from level.yml in the future.
Everything was tested by me.

UPDATE:
Well, I have made use of the PlayerXpGainEvent event to update the experience gained in the course of the game. On the other hand a new event is created "PlayerMoneyGainEvent" that fulfills the same function, to update the money earned. This event is called in each method that money is given to the player.

Is there a better way to store the experience and money earned instead of a static hashmap?

Also remove odd spaces, improve some variable names to follow Java conventions and make the code more homogeneous with respect to other variables.